requirements.txt
# core
fastapi>=0.118.0,<0.130.0
asyncpg>=0.30.0,<0.31.0
psycopg2-binary>=2.9.9,<3.0.0
sqlalchemy>=2.0.43,<2.1.0
# secodnary
python-jose[cryptography]>=3.5.0,<4.0.0
uvicorn[standard]>=0.37.0,<0.40.0
pydantic[email]>=2.11.9,<2.12.0
pydantic-settings>=2.11.0,<2.12.0
alembic>=1.16.5,<1.17.0
passlib[cryptography]>=1.7.4,<2.0.0
# distributed
redis>=6.4.0,<7.0.0
celery>=5.5.3,<6.0.0
itsdangerous>=2.2.0,<3.0.0